+++ core-ng/src/main/java/core/framework/api/http/ContentType.java
@@ -2,8 +2,11 @@ package core.framework.api.http;
 
 import core.framework.api.util.ASCII;
 import core.framework.api.util.Charsets;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.nio.charset.Charset;
+import java.nio.charset.UnsupportedCharsetException;
 import java.util.Optional;
 
 /**
@@ -18,6 +21,8 @@ public final class ContentType {
     public static final ContentType APPLICATION_JAVASCRIPT = create("application/javascript", Charsets.UTF_8);
     public static final ContentType APPLICATION_OCTET_STREAM = create("application/octet-stream", null);
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(ContentType.class);
+
     // only cover common case, assume pattern is "media-type; charset=" or "multipart/form-data; boundary="
     public static ContentType parse(String contentType) {
         String mediaType = contentType;
@@ -28,11 +33,21 @@ public final class ContentType {
             mediaType = contentType.substring(0, firstSemicolon);
 
             int charsetIndex = contentType.indexOf("charset=", firstSemicolon + 1);
-            if (charsetIndex > 0)
-                charset = Charset.forName(contentType.substring(charsetIndex + 8));
+            if (charsetIndex > 0) {
+                charset = parseCharset(contentType.substring(charsetIndex + 8));
+            }
         }
 
         return new ContentType(contentType, mediaType, charset);
+    }
+
+    private static Charset parseCharset(String charset) {
+        try {
+            return Charset.forName(charset);
+        } catch (UnsupportedCharsetException e) {
+            LOGGER.warn("ignore unsupported charset, charset={}", charset);
+            return null;
+        }
     }
 
     public static ContentType create(String mediaType, Charset charset) {