+++ core-ng/src/main/java/core/framework/impl/web/session/SessionImpl.java
@@ -33,6 +33,7 @@ public class SessionImpl implements Session {
     @Override
     public void remove(String key) {
         data.remove(key);
+        changed = true;
     }
 
     @Override