+++ core-ng/src/main/java/core/framework/impl/web/session/SessionManager.java
@@ -17,17 +17,18 @@ import java.util.UUID;
  */
 public class SessionManager {
     private final Logger logger = LoggerFactory.getLogger(SessionManager.class);
-    private SessionStore sessionStore;
-    private Duration sessionTimeout = Duration.ofMinutes(20);
+    public String cookieDomain;     // default to be null, which will be current host
+    private Duration timeout = Duration.ofMinutes(20);
+    private SessionStore store;
 
     public Session load(RequestImpl request) {
-        if (sessionStore == null) return null;  // session store is not initialized
+        if (store == null) return null;  // session store is not initialized
 
         logger.debug("load http session");
         SessionImpl session = new SessionImpl();
 
         request.cookie("SessionId").ifPresent(sessionId -> {
-            Map<String, String> sessionData = sessionStore.getAndRefresh(sessionId, sessionTimeout);
+            Map<String, String> sessionData = store.getAndRefresh(sessionId, timeout);
             if (sessionData != null) {
                 session.id = sessionId;
                 session.data.putAll(sessionData);
@@ -43,7 +44,7 @@ public class SessionManager {
 
         logger.debug("save http session");
         if (session.invalidated && session.id != null) {
-            sessionStore.clear(session.id);
+            store.clear(session.id);
             CookieImpl cookie = sessionCookie(request.scheme());
             cookie.setMaxAge(0);
             cookie.setValue("");
@@ -56,22 +57,24 @@ public class SessionManager {
                 cookie.setValue(session.id);
                 exchange.setResponseCookie(cookie);
             }
-            sessionStore.save(session.id, session.data, sessionTimeout);
+            store.save(session.id, session.data, timeout);
         }
     }
 
     public void sessionStore(SessionStore sessionStore) {
-        if (this.sessionStore != null)
-            throw Exceptions.error("session store is already configured, previous={}", this.sessionStore);
-        this.sessionStore = sessionStore;
+        if (this.store != null)
+            throw Exceptions.error("session store is already configured, previous={}", this.store);
+        this.store = sessionStore;
     }
 
-    public void sessionTimeout(Duration sessionTimeout) {
-        this.sessionTimeout = sessionTimeout;
+    public void timeout(Duration timeout) {
+        if (timeout == null) throw Exceptions.error("timeout must not be null");
+        this.timeout = timeout;
     }
 
     private CookieImpl sessionCookie(String scheme) {
         CookieImpl cookie = new CookieImpl("SessionId");
+        cookie.setDomain(cookieDomain);
         cookie.setPath("/");
         cookie.setSecure("https".equals(scheme));
         cookie.setHttpOnly(true);