+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 * faster synchronous rpc
 
 ## Change log
+### 4.6.1 (6/22/2016)
+* httpclient: ContentType.parse supports to ignore illegal charset
+* session: support to configure cookie domain
+
 ### 4.6.0 (6/20/2016 - 6/21/2016)
 * mongo: refine mongo decoding logic and error message
 * http: fix webContext should init at beginning, to make it available in 404 error handler